+++ airbyte-connector-builder-server/Dockerfile
@@ -3,7 +3,7 @@ FROM python:3.9-slim as base
 RUN apt-get upgrade \
     && pip install --upgrade pip
 
-WORKDIR /home/connector-builder-server
+WORKDIR /home/connector-builder
 COPY . ./
 
 RUN pip install --no-cache-dir .
@@ -11,5 +11,5 @@ RUN pip install --no-cache-dir .
 ENTRYPOINT ["uvicorn", "connector_builder.entrypoint:app", "--host", "0.0.0.0", "--port", "80"]
 
 LABEL io.airbyte.version=0.40.15
-LABEL io.airbyte.name=airbyte/connector-builder-server
+LABEL io.airbyte.name=airbyte/connector-builder
