+++ airbyte-connector-builder-server/connector_builder/entrypoint.py
@@ -2,29 +2,14 @@
 # Copyright (c) 2022 Airbyte, Inc., all rights reserved.
 #
 
-from typing import Union
-
+from connector_builder.generated.apis.default_api_interface import initialize_router
+from connector_builder.impl.default_api import DefaultApiImpl
 from fastapi import FastAPI
-from fastapi.middleware.cors import CORSMiddleware
 
-app = FastAPI()
-
-origins = ["*"]
-
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=origins,
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
+app = FastAPI(
+    title="Connector Builder Server API",
+    description="Connector Builder Server API ",
+    version="1.0.0",
 )
 
-
-@app.get("/")
-def read_root():
-    return {"Hello": "fastapi"}
-
-
-@app.get("/items/{item_id}")
-def read_item(item_id: int, q: Union[str, None] = None):
-    return {"item_id": item_id, "q": q}
+app.include_router(initialize_router(DefaultApiImpl()))
