+++ airbyte-connector-builder-server/openapi/generator_config.yaml
