+++ airbyte-connector-builder-server/openapi/templates/api_interfaces.mustache
