--- airbyte-connector-builder-server/run.sh
@@ -1,1 +1,0 @@
-uvicorn connector_builder.entrypoint:app --host 0.0.0.0 --port 80
