+++ airbyte-connector-builder-server/setup.py
@@ -13,7 +13,7 @@ HERE = pathlib.Path(__file__).parent
 README = (HERE / "README.md").read_text()
 
 setup(
-    name="connector-builder-server",
+    name="connector-builder",
     version="0.0.1",
     description="",
     long_description=README,
@@ -41,24 +41,9 @@ setup(
     },
     packages=find_packages(exclude=("unit_tests", "integration_tests", "docs")),
     package_data={},
-    install_requires=[
-        "click~=8.0.3",
-        "fastapi>=0.68.0,<0.69.0",
-        "pydantic>=1.8.0,<2.0.0",
-        "uvicorn>=0.15.0,<0.16.0",
-        # f"airbyte_connector_builder @ file://{os.getcwd()}/build/airbyte_connector_builder", # FIXME: This should point to the server interface generated by the openAPI generator
-        "jinja2~=3.0.3",
-        "deepdiff~=5.7.0",
-        "pyyaml~=6.0",
-        "analytics-python~=1.4.0",
-        "python-slugify~=6.1.2",
-    ],
+    install_requires=["fastapi", "uvicorn"],
     python_requires=">=3.9.11",
     extras_require={
         "tests": ["MyPy~=0.812", "pytest~=6.2.5", "pytest-cov", "pytest-mock", "pytest-recording", "requests-mock", "pre-commit"],
-        "sphinx-docs": [
-            "Sphinx~=4.2",
-            "sphinx-rtd-theme~=1.0",
-        ],
     },
 )
