--- airbyte-connector-builder/.coveragerc
@@ -1,3 +1,0 @@
-[report]
-# show lines missing coverage
-show_missing = true
