--- airbyte-connector-builder/.dockerignore
@@ -1,4 +1,0 @@
-build
-!build/airbyte_api_client
-.venv
-connector_builder.egg-info
