--- airbyte-connector-builder/.gitignore
@@ -1,3 +1,0 @@
-.coverage
-.venv
-state_*.yaml
