--- airbyte-connector-builder/build.gradle
@@ -1,45 +1,0 @@
-import org.openapitools.generator.gradle.plugin.tasks.GenerateTask
-
-plugins {
-    id "org.openapi.generator" version "5.3.1"
-    id 'airbyte-python'
-    id 'airbyte-docker'
-}
-
-airbytePython {
-    moduleDirectory 'connector_builder'
-}
-
-task generateOpenApiPythonServer(type: GenerateTask){
-    outputs.upToDateWhen { false }
-
-    def generatedCodeDir = "$buildDir/server"
-    inputSpec = "$rootDir.absolutePath/airbyte-connector-builder/src/main/openapi/openapi.yaml"
-    outputDir = generatedCodeDir
-
-    generatorName = "python-fastapi"
-    configFile = "$projectDir/openapi/generator_config.yaml"
-    templateDir = "$projectDir/openapi/templates"
-    packageName = "connector_builder.generated"
-
-    // After we generate, we're only interested in the API declaration and the generated pydantic models.
-    // So we copy those from the build/ directory
-    doLast {
-        def sourceDir = "$generatedCodeDir/src/connector_builder/generated/"
-        def targetDir = "$projectDir/connector_builder/generated"
-        mkdir targetDir
-        copy {
-            from "$sourceDir/apis"
-            include "*_interface.py", "__init__.py"
-            into "$targetDir/apis"
-        }
-        copy {
-            from "$sourceDir/models"
-            include "*.py"
-            into "$targetDir/models"
-        }
-    }
-}
-
-project.build.dependsOn(generateOpenApiPythonServer)
-
