--- airbyte-connector-builder/connector_builder/__init__.py
@@ -1,3 +1,0 @@
-#
-# Copyright (c) 2022 Airbyte, Inc., all rights reserved.
-#
