--- airbyte-connector-builder/pytest.ini
@@ -1,7 +1,0 @@
-[pytest]
-log_cli = 1
-log_cli_level = INFO
-log_cli_format = %(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)
-log_cli_date_format=%Y-%m-%d %H:%M:%S
-markers =
-    integration: marks tests as integration test (deselect with '-m "not integration"')
